### 2021-04-13 #38 @mkandziora
PhylUp: updating alignments with custom taxon sampling
## Package Info
Submitting Author: Martha Kandziora (@mkandziora)
Repository Link (if existing):   https://github.com/mkandziora/PhylUp

```
PhylUp - a python package to automatically update alignments
using local sequences or a local Genbank database
while controlling for the number of sequences per OTU.
```

- If you selected Unsure/Other, briefly explain:

## How Can We Help?

- Indicate how we can help you prepare your package (select all that apply):
	- [ ] Creating a test suite
	- [x] Enabling Continuous Integration
	- [ ] Adding a specific feature
	- [ ] Improving Documentation
	- [x] General Assistance
	- [ ] Other

- Explain briefly below if there's something specific you'd like help with:
  I am quite new to programming and package development and would need some general help to comply with the requirements for the pyOpenSci submission, e.g.  (re-)enabling automatic testing, packaging, readthedocs

  TRAVIS CI is implemented, but I think due to a wrong setup I ran out of 'credit'. The package was developed without funding and my current institution has no TRAVIS subscription.
