### 2021-04-13 df556b5
fix class assignment
