+++ update_databases.py
@@ -8,9 +8,9 @@ conf = config.ConfigObj(configfi, workdir, interactive=True)
 
 db_updater._download_localblastdb(conf)
 
-ncbi_data_parser.Parser(names_file=conf.ncbi_parser_names_fn,
+ncbi_parser = ncbi_data_parser.Parser(names_file=conf.ncbi_parser_names_fn,
                         nodes_file=conf.ncbi_parser_nodes_fn)
 
 
 
-ncbi_data_parser._download_ncbi_parser(conf)
+ncbi_parser._download_ncbi_parser(conf)
