+++ README.md
@@ -9,18 +9,17 @@ A preprint is available on [biorxiv](https://www.biorxiv.org/content/10.1101/202
     
 ## Short introduction
 
-PhylUp is a command-line program written in python3 to automatically update alignments and phylogenies with a focus on different sampling strategies.
+PhylUp is a command-line program written in python3 to automatically build alignments from a single initial seed sequence or update alignments with a focus on different taxonomic sampling strategies.
 As input it needs a alignment or a single sequence (and if available a phylogeny) and 
 a file with the information about the sequence names and the corresponding species names. 
 PhylUp will take every input sequence and blasts it against the ncbi GenBank database. 
-Sequences that are similar to the input sequence will be added to the alignment, 
+Sequences that are similar to the input sequence will be added, 
 if they are a different taxon and/or they are longer than existing sequences.
 They are then filtered to user settings provided in the configuration file.
-Newly found and filtered sequences will be blasted again until no new sequences were found.
-Finally, it will place the newly found sequences into the alignment and if enabled to update phylogenies.
+Newly found and filtered sequences will be blasted again until no new sequences can be found.
+Finally, it will place the newly found sequences into the alignment and if enabled update phylogenies.
 
-After the single-gene datasets are updated, the data can be concatenated. 
-The tool decides randomly which sequences to combine if there are more than a single sequence for a taxon in one of the alignments.
+After the single-gene datasets are updated, the data can be [concatenated](https://github.com/mkandziora/phylogenetic_concatenation). 
 
 
 ## Tutorial
