+++ PhylUp/wrapper.py
@@ -45,9 +45,9 @@ def run_multiple(data, confs, end, overlap_folder=None, first_locus=False):
                         files['trfn'] = "{}/updt_tre.tre".format(files['workdir'])
                     files['seqaln'] = "{}/updt_aln.fasta".format(files['workdir'])
             conf = config.ConfigObj(conffi, files['workdir'], interactive=False)
-            if conf.preferred_taxa == True:
-                assert overlap_folder != None
-                if end == None or end >=1:
+            if conf.preferred_taxa is True:
+                assert overlap_folder is not None
+                if end is None or end >= 1:
                     end = 1
 
             test = phyl_up.PhylogeneticUpdater(files['idtospn'], files['seqaln'], files['mattype'], files['trfn'],
@@ -55,20 +55,21 @@ def run_multiple(data, confs, end, overlap_folder=None, first_locus=False):
 
             test.run(status_end=end)
 
-        if conf.preferred_taxa == True:
+        if conf.preferred_taxa is True:
             print('preferred taxa: select')
             found_taxa_list = {}
             for locus in data.keys():
                 files = data[locus]
                 found_taxa_list[locus] = "{}/found_taxa.csv".format(files['workdir'])
             mrca = test.set_mrca(test.config.mrca_input)
-            if conf.preferred_taxa_fn == None:
+            if conf.preferred_taxa_fn is None:
                 if not os.path.exists(overlap_folder):
                     os.mkdir(overlap_folder)
                 conf.preferred_taxa_fn = os.path.join(overlap_folder, 'overlap{}.csv'.format(mrca))
                 print(conf.preferred_taxa_fn)
-                assert conf.preferred_taxa_fn != None, conf.preferred_taxa_fn
-                phylogenetic_helpers.make_preferred_taxon_list(found_taxa_list, conf.preferred_taxa_fn, overlap_complete=True)
+                assert conf.preferred_taxa_fn is not None, conf.preferred_taxa_fn
+                phylogenetic_helpers.make_preferred_taxon_list(found_taxa_list, conf.preferred_taxa_fn,
+                                                               overlap_complete=True)
             else:
                 assert os.path.exists(conf.preferred_taxa_fn), conf.preferred_taxa_fn
 
@@ -83,17 +84,17 @@ def run_multiple(data, confs, end, overlap_folder=None, first_locus=False):
                         files['seqaln'] = "{}/updt_aln.fasta".format(files['workdir'])
 
                 conf = config.ConfigObj(conffi, files['workdir'], interactive=False)
-                if conf.preferred_taxa_fn == None:
+                if conf.preferred_taxa_fn is None:
                     conf.preferred_taxa_fn = os.path.join(overlap_folder,  'overlap{}.csv'.format(mrca))
                     print(conf.preferred_taxa_fn)
-                    assert conf.preferred_taxa_fn != None, conf.preferred_taxa_fn
+                    assert conf.preferred_taxa_fn is not None, conf.preferred_taxa_fn
 
                 test = phyl_up.PhylogeneticUpdater(files['idtospn'], files['seqaln'], files['mattype'], files['trfn'],
                                                    files['schema_trf'], conf, ignore_acc_list=files['ignore_acc_list'])
                 test.run(status_end=end)
 
                 # new implementation - restricts overlap to species from first locus.
-                if first_locus == True:
+                if first_locus is True:
                     tf = test.table['status'] >= 0
                     print(tf)
                     added_taxa = test.table[tf]
