### 2021-04-27 712cbe7
new examples
also mentioned in README
