### 2021-04-27 #38 @NickleDave
Hi @mkandziora thank you for your interest in contributing to PyOpenSci, and I'm sorry we haven't gotten back to you sooner.

I would love to get you help with developing `phylup`.

For CI, I would suggest switching to GitHub Actions.
https://docs.github.com/en/actions/learn-github-actions/migrating-from-travis-ci-to-github-actions
I think that as long as your package is public then CI should be free for you.
If you get stuck, just let us know.

If you have other specific questions, it might be better if we chat on our Discourse forum:
https://pyopensci.discourse.group/
Feel free to tag me in any questions you ask so I don't miss them
