+++ .github/workflows/main.yml
@@ -1,0 +1,35 @@
+name: PhylUp integration
+
+on:
+  push:
+    branches:
+      - main
+      - dev
+
+jobs:
+  run_python:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/setup-python@v2
+        with:
+          python-version: '3.7'
+          architecture: 'x64'
+
+
+      -uses: mkandziora/PhylUp
+
+      - name: Install dependencies
+            run: |
+              python -m pip install --upgrade pip
+              pip install flake8 pytest pytest-cov pytest-xdist codecov
+              if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+
+      - name: Test with pytest
+          run: |
+            pytest --cov=PhylUp tests/test_*
+
+      - name: upload coverage
+          run: |
+            codecov
+            curl -s https://codecov.io/bash | bash
+
