+++ .github/workflows/main.yml
@@ -14,20 +14,15 @@ jobs:
         with:
           python-version: '3.7'
           architecture: 'x64'
-
-
       -uses: mkandziora/PhylUp
-
       - name: Install dependencies
             run: |
               python -m pip install --upgrade pip
               pip install flake8 pytest pytest-cov pytest-xdist codecov
               if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-
       - name: Test with pytest
           run: |
             pytest --cov=PhylUp tests/test_*
-
       - name: upload coverage
           run: |
             codecov
