+++ .github/workflows/main.yml
@@ -16,7 +16,7 @@ jobs:
           architecture: 'x64'
       - uses: mkandziora/PhylUp
       - name: Install dependencies
-            run: |
+          run: |
               python -m pip install --upgrade pip
               pip install flake8 pytest pytest-cov pytest-xdist codecov
               if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
