+++ .github/workflows/main.yml
@@ -10,11 +10,10 @@ jobs:
   run_python:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/setup-python@v2
+      - uses: mkandziora/PhylUp
         with:
           python-version: '3.7'
           architecture: 'x64'
-      - uses: mkandziora/PhylUp
       - name: Install dependencies
           run: |
               python -m pip install --upgrade pip
