+++ .github/workflows/main.yml
@@ -10,20 +10,21 @@ jobs:
   run_python:
     runs-on: ubuntu-latest
     steps:
-      - uses: mkandziora/PhylUp
+      - uses: actions/setup-python@v2
         with:
           python-version: '3.7'
           architecture: 'x64'
+      - uses: mkandziora/PhylUp
       - name: Install dependencies
-          run: |
+        run: |
               python -m pip install --upgrade pip
-              pip install flake8 pytest pytest-cov pytest-xdist codecov
+              pip install pytest pytest-cov pytest-xdist codecov
               if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test with pytest
-          run: |
+        run: |
             pytest --cov=PhylUp tests/test_*
       - name: upload coverage
-          run: |
+        run: |
             codecov
             curl -s https://codecov.io/bash | bash
 
