+++ .github/workflows/main.yml
@@ -14,7 +14,7 @@ jobs:
         with:
           python-version: '3.7'
           architecture: 'x64'
-      - uses: mkandziora/PhylUp
+      - uses: mkandziora/PhylUp@dev
       - name: Install dependencies
         run: |
               python -m pip install --upgrade pip
