+++ .github/workflows/main.yml
@@ -14,6 +14,7 @@ jobs:
         with:
           python-version: '3.7'
           architecture: 'x64'
+      - uses: actions/checkout@v2
       - uses: mkandziora/PhylUp@dev
       - name: Install dependencies
         run: |
