+++ .github/workflows/main.yml
@@ -10,17 +10,17 @@ jobs:
   run_python:
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
         with:
           python-version: '3.7'
           architecture: 'x64'
-      - uses: actions/checkout@v2
       - uses: mkandziora/PhylUp@dev
       - name: Install dependencies
         run: |
               python -m pip install --upgrade pip
               pip install pytest pytest-cov pytest-xdist codecov
-              if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+              pip install -r requirements.txt
       - name: Test with pytest
         run: |
             pytest --cov=PhylUp tests/test_*
