+++ .github/workflows/main.yml
@@ -15,7 +15,7 @@ jobs:
         with:
           python-version: '3.7'
           architecture: 'x64'
-      - uses: mkandziora/PhylUp@dev
+      #- uses: mkandziora/PhylUp@dev
       - name: Install dependencies
         run: |
               python -m pip install --upgrade pip
