+++ .github/workflows/main.yml
@@ -21,6 +21,12 @@ jobs:
               python -m pip install --upgrade pip
               pip install pytest pytest-cov pytest-xdist codecov
               pip install -r requirements.txt
+              python setup.py install
+      - name: download data
+        run: |
+            wget 'https://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz'
+            gunzip  -cd taxdump.tar.gz | (tar xvf - names.dmp nodes.dmp)
+            mv *.dmp ./data/
       - name: Test with pytest
         run: |
             pytest --cov=PhylUp tests/test_*
