+++ .github/workflows/main.yml
@@ -15,12 +15,14 @@ jobs:
         with:
           python-version: '3.7'
           architecture: 'x64'
-      - uses: conda-incubator/setup-miniconda@v2
-          with:
-            installer-url: https://github.com/conda-forge/miniforge/releases/download/4.8.3-2/Miniforge-pypy3-4.8.3-2-Linux-x86_64.sh
-            allow-softlinks: true
-            show-channel-urls: true
-            use-only-tar-bz2: true
+#      - uses: conda-incubator/setup-miniconda@v2
+#        with:
+#          installer-url: https://github.com/conda-forge/miniforge/releases/download/4.8.3-2/Miniforge-pypy3-4.8.3-2-Linux-x86_64.sh
+#          allow-softlinks: true
+#          show-channel-urls: true
+#          use-only-tar-bz2: true
+          #conda install -c bioconda blast
+
       #- uses: mkandziora/PhylUp@dev
       - name: Install dependencies
         run: |
@@ -28,7 +30,13 @@ jobs:
               pip install pytest pytest-cov pytest-xdist codecov
               pip install -r requirements.txt
               python setup.py install
-              conda install -c bioconda blast
+
+              wget ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/2.9.0/ncbi-blast-2.9.0+-x64-linux.tar.gz
+              tar -xvf ncbi-blast-2.9.0+-x64-linux.tar.gz
+              cd ncbi-blast-2.9.0+/bin/
+              echo export PATH="$PATH:$(pwd)" >> ~/.bashrc
+              cd ..
+              cd ..
       - name: download data
         run: |
             wget 'https://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz'
