### 2021-04-28 31168f7
add path to bash
