+++ .github/workflows/main.yml
@@ -34,7 +34,7 @@ jobs:
               wget ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/2.9.0/ncbi-blast-2.9.0+-x64-linux.tar.gz
               tar -xvf ncbi-blast-2.9.0+-x64-linux.tar.gz
               cd ncbi-blast-2.9.0+/bin/
-              echo export PATH="$PATH:$(pwd)" >> ~/.bashrc
+              echo PATH="$PATH:$(pwd)" >> $GITHUB_PATH
               cd ..
               cd ..
       - name: download data
@@ -46,7 +46,12 @@ jobs:
             mkdir tests/output/test_runs
       - name: Test with pytest
         run: |
-            pytest --cov=PhylUp tests/test_*
+            pytest --cov=PhylUp tests/test_aln_updater.py
+            pytest --cov=PhylUp tests/test_filter_otu.py
+            pytest --cov=PhylUp tests/test_phylhelpers.py
+            pytest --cov=PhylUp tests/test_reconcile.py
+            pytest --cov=PhylUp tests/test_standard.py
+
       - name: upload coverage
         run: |
             codecov
