+++ .github/workflows/main.yml
@@ -62,6 +62,6 @@ jobs:
 #            curl -s https://codecov.io/bash | bash
 
       - name: Upload coverage to Codecov
-          uses: codecov/codecov-action@v1
+        uses: codecov/codecov-action@v1
 
 
