+++ tests/test_phylogen_updater.py
@@ -28,9 +28,9 @@ def test_remove_short_fromaln():
     aln0 = deepcopy(phylogenetic_helpers.read_in_aln(seqaln, mattype).taxon_namespace)
     test = phyl_up.PhylogeneticUpdater(id_to_spn, seqaln, mattype, trfn, schema_trf, conf)
     alnbefore = deepcopy(test.aln.taxon_namespace)
+    phylogenetic_helpers.write_aln(test.aln, test.config.workdir)
 
     aln_upd = phylogen_updater.AlnUpdater(conf, test.aln, test.table, 2, None)
-    phylogenetic_helpers.write_aln(aln_upd.aln, test.config.workdir)
 
     alnafter = deepcopy(aln_upd.aln.taxon_namespace)
 
