### 2021-04-28 219b9ed
delete short from aln if not updated
