+++ tests/test_standard.py
@@ -42,6 +42,38 @@ def test_standard_run():
    # shutil.copyfile('data/tiny_test_example/updt_tre.tre', os.path.join(workdir, 'updt_tre.tre'))
 
     test = phyl_up.PhylogeneticUpdater(id_to_spn, seqaln, mattype, trfn, schema_trf, conf, ignore_acc_list=None)
+    test.run()
+
+def xtest_standard_run():
+
+    workdir = "tests/output/test_runs"  # working directory
+    trfn = "data/tiny_test_example/test.tre"  # phylogeny
+    schema_trf = "newick"  # format of phylogeny
+    id_to_spn = "data/tiny_test_example/test_nicespl.csv"  # tab-delimited file where tip names correspond to ncbi names
+    seqaln = "data/tiny_test_example/test.fas"  # alignment
+    mattype = "fasta"  # format of alignment
+    configfi = "data/localblast_test.config"
+
+    conf = config.ConfigObj(configfi, workdir, interactive=False)
+    conf.threshold = 2
+    conf.blast_folder = os.path.abspath("./data/blast_for_tests")
+    conf.identical_seqs = False
+
+
+
+    if not os.path.exists(workdir):
+        os.rename(workdir)
+    if not os.path.exists(workdir):
+        os.mkdir(workdir)
+    tmp_folder = os.path.join(workdir, 'tmp')
+    if not os.path.exists(tmp_folder):
+        os.mkdir(tmp_folder)
+    # call(['cp', '-a', 'data/tmp_for_test/', tmp_folder])
+    copy_tree('data/tmp_for_test/', tmp_folder)
+   # shutil.copyfile('data/tiny_test_example/updt_aln.fasta', os.path.join(workdir, 'updt_aln.fasta'))
+   # shutil.copyfile('data/tiny_test_example/updt_tre.tre', os.path.join(workdir, 'updt_tre.tre'))
+
+    test = phyl_up.PhylogeneticUpdater(id_to_spn, seqaln, mattype, trfn, schema_trf, conf, ignore_acc_list=None)
     #test.run()
 
    # update the data:
