### 2021-04-29 ee38793
add helper scripts
