### 2021-04-30 7633e02
update docstrings
