+++ PhylUp/db_updater.py
@@ -7,6 +7,8 @@ Package to automatically generate alignments (or update alignments and phylogeni
 using local sequences or a local Genbank database
 while controlling for the number of sequences per OTU and taxonomic rank.
 
+This submodule is used to update needed ncbi databases.  Needs user input, so do not submit it as a job on a cluster.
+
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 3 of the License, or
