### 2021-04-30 3dba93b
update helper scripts - argparse
