+++ data/helper_scripts/make_spn_file.py
@@ -1,6 +1,6 @@
 """
-Copyright (C) 2019  Martha Kandziora
-martha.kandziora@yahoo.com
+Martha Kandziora
+martha.kandziora@mailbox.org
 
 make a spn name file from alignment - test if all names are accepted using check_spn.py
 """
@@ -10,17 +10,10 @@ import csv
 import os
 from Bio import SeqIO
 
-
 # create variables that can be entered in the command line
-parser = argparse.ArgumentParser(usage='''
-check_spn.py [<args>]
-The check_spn.py arguments is:
-  -aln   STR     path to file with alignment (required)
-''')
-
-parser.add_argument('-aln', required = True, help = 'path to file with alignment')
-
-
+parser=argparse.ArgumentParser(description='make a spn name file from alignment - '
+                                           'test if all names are accepted using check_spn.py')
+parser.add_argument('-aln', required=True, help='path to file with alignment')
 args = parser.parse_args()
 
 ncbi_parser = ncbi_data_parser.Parser(names_file='./data/names.dmp', nodes_file="./data/nodes.dmp")
