+++ tests/tests_setup.py
@@ -1,9 +1,12 @@
 import os
-from shutil import copyfile
+from shutil import copyfile, rmtree
 
 
 src = "./data/data_for_tests/updt_aln.fasta"
 dst = "./tests/output/test_runs/updt_aln.fasta"
+if os.path.exists("./tests/output/"):
+    rmtree("./tests/output/")
+    
 os.mkdir("./tests/output/")
 os.mkdir("./tests/output/test_runs/")
 copyfile(src, dst)
