+++ .travis.yml
@@ -1,183 +1,0 @@
-dist: trusty
-sudo: required
-language: python
-python:
-  - "3.6"
-compiler:
-  - gcc
-addons:
-  apt:
-    sources:
-      - ubuntu-toolchain-r-test
-    packages:
-      - gcc-7
-      - g++-7
-      - cmake
-
-before_install:
-  - sudo apt-get update
-
-  # Link gcc-6 and g++-6 to their standard commands
-  - sudo ln -s /usr/bin/gcc-7 /usr/local/bin/gcc
-  - sudo ln -s /usr/bin/g++-7 /usr/local/bin/g++
-  # Export CC and CXX to tell cmake which compiler to use
-  - export CC=/usr/bin/gcc-7
-  - export CXX=/usr/bin/g++-7
-  # Check versions of gcc, g++ and cmake
-  - gcc -v && g++ -v && cmake --version
-  
-  
-  # Run your build commands next
-
-  # use raxml-ng from bioconda - there are compilation errors...
-  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
-  - chmod +x miniconda.sh
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q conda
-  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
-  - source activate test-environment
-  #- pip install --force-reinstall numpy>= 1.15.4
-  - conda install -c bioconda raxml-ng
-
-
-#  # get newer cmake version for raxml-ng
-#  # first we create a directory for the CMake binaries
-#  - DEPS_DIR="${TRAVIS_BUILD_DIR}/deps"
-#  - mkdir ${DEPS_DIR} && cd ${DEPS_DIR}
-#  # we use wget to fetch the cmake binaries
-#  - travis_retry wget --no-check-certificate https://cmake.org/files/v3.3/cmake-3.3.2-Linux-x86_64.tar.gz
-#  # this is optional, but useful:
-#  # do a quick md5 check to ensure that the archive we downloaded did not get compromised
-#  - echo "f3546812c11ce7f5d64dc132a566b749 *cmake-3.3.2-Linux-x86_64.tar.gz" > cmake_md5.txt
-#  - md5sum -c cmake_md5.txt
-#  # extract the binaries; the output here is quite lengthy,
-#  # so we swallow it to not clutter up the travis console
-#  - tar -xvf cmake-3.3.2-Linux-x86_64.tar.gz > /dev/null
-#  - mv cmake-3.3.2-Linux-x86_64 cmake-install
-#  # add both the top-level directory and the bin directory from the archive
-#  # to the system PATH. By adding it to the front of the path we hide the
-#  # preinstalled CMake with our own.
-#  - PATH=${DEPS_DIR}/cmake-install:${DEPS_DIR}/cmake-install/bin:$PATH
-#  # don't forget to switch back to the main build directory once you are done
-#  - cd ${TRAVIS_BUILD_DIR}
-
-#  # raxml-ng
-#  - mkdir RAxML-ng
-#  - cd RAxML-ng
-#  - wget https://github.com/amkozlov/raxml-ng/releases/download/0.9.0/raxml-ng_v0.9.0_linux_x86_64_MPI.zip
[…]
