### 2021-04-30 #38 @lwasser
oh hi! @NickleDave agreed with the idea of moving this discussion to discourse. let's talk about slack as well this week. i think it will be easiest :) @mkandziora welcome to pyopensci!! we are happy help you however we can.  i do suggest  github actions as well.
