### 2021-05-02 #38 @mkandziora
Hi  @NickleDave and @lwasser. Thank you very much for the suggestion on GitHub Actions. I did not know about this option - very convenient and I successfully managed to migrate.

I am currently working on moving my documentation from the GitHub wiki to a readthedocs - but documentation is there.

I think I now fulfill all requirements for submission of the code for a pyOpenScience review. Shall I officially start a review request or do we continue here?
