### 2021-05-15 #38 @NickleDave
Hi again @mkandziora -- glad we could help.

Also great to hear you are moving to readthedocs.

Sorry for the delayed response.
Not sure if you saw in our forums or other reviews but we have been on hiatus.
@lwasser was recently awarded funding so that will be changing near the end of summer.

Long story short:
the current  are meeting Monday to sync up and will have a better idea of where we are with reviews after then.
There are a couple of potential reviews that were raised before yours, and to be fair to them we should make sure we address those first.

Let me get back to you early next week and update you on where we're at.
