### 2021-05-20 #38 @NickleDave
@mkandziora I am going to close this issue.

> I think I now fulfill all requirements for submission of the code for a pyOpenScience review. Shall I officially start a review request or do we continue here?

Please do go ahead and open a new issue requesting review. I can act as editor.

It looks to me like `PhylUp` is within scope. I think you might also want to add "data munging" as a category the package falls under. We don't really have a category (currently) that denotes the pipeline/workflow aspects of what `PhylUp` does but I think "mugning" is close. I'm a bit worried there might be a little overlap with existing packages, but my initial take is that what `PhylUp` does is different enough that it's not the case. I will need to get input from people with more bioinformatics expertise. Let's discuss that on an issue about review though.

Thank you!
